requirements.txt
fastapi
uvicorn
python-dotenv
glob2
langchain
langchain-community
chromadb
openai
watchdog
